refactor: slim down skill scanner to pattern-only baseline
- Remove yara_analyzer, consistency_analyzer and all YARA rule files
  so they can be introduced in dedicated follow-up branches/PRs.
- Move SkillScanError / scan_skill_directory from skill_scan.py into
  skill_scanner/__init__.py (mirrors tool_guard layout).
- Delete the now-redundant skill_scan.py shim.
- Clean security/__init__.py to be import-free.
- Strip yara/consistency-specific fields from ScanPolicy and
  default_policy.yaml.
- Update import paths in skills_hub.py and skills_manager.py.

Made-with: Cursor
